refactor: reorganize post API types for better structure and maintainability
